+++ tests/golden/syscall-hello-elf-o1-c/source.c
@@ -22,7 +22,7 @@ int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t ar
     int64_t v0 = 0LL;
     int64_t v1 = 0LL;
     int64_t v2 = 0LL;
-    int64_t v3 = 0LL;
+    void * v3 = ((void *)(0LL));
     int64_t v4 = 0LL;
     int64_t v5 = 0LL;
     int64_t v6 = 0LL;
@@ -38,8 +38,8 @@ int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t ar
     int64_t v16 = 0LL;
 
     v1 = (v0 - 8LL);
-    v3 = (v2 + 16336LL);
-    v4 = (*((int64_t *)(v3)));
+    v3 = ((void *)((v2 + 16336LL)));
+    v4 = (*((int64_t *)(((int64_t)(v3)))));
     v5 = v4;
     v6 = (v5 & v5);
     v7 = (v6 == 0LL);
@@ -104,13 +104,13 @@ void _start(int64_t arg0, int64_t arg1, int64_t arg2) {
     int64_t v1 = 0LL;
     int64_t v2 = arg1;
     int64_t v3 = 0LL;
-    int64_t v4 = 0LL;
+    void * v4 = ((void *)(0LL));
     int64_t v5 = 0LL;
-    int64_t v6 = 0LL;
+    void * v6 = ((void *)(0LL));
     int64_t v7 = 0LL;
-    int64_t v8 = 0LL;
-    int64_t v9 = 0LL;
-    int64_t v10 = 0LL;
+    void * v8 = ((void *)(0LL));
+    void * v9 = ((void *)(0LL));
+    void * v10 = ((void *)(0LL));
     int64_t v11 = 0LL;
     int64_t v12 = 0LL;
     int64_t v13 = arg2;
@@ -124,20 +124,20 @@ void _start(int64_t arg0, int64_t arg1, int64_t arg2) {
 
     v1 = (v0 ^ v0);
     v3 = v2;
-    v5 = (*((int64_t *)(v4)));
-    v6 = (v4 + 8LL);
+    v5 = (*((int64_t *)(((int64_t)(v4)))));
+    v6 = ((void *)((((int64_t)(v4)) + 8LL)));
     v7 = v5;
-    v8 = v6;
-    v9 = (v6 & -16LL);
-    v10 = (v9 - 8LL);
-    *((int64_t *)(v10)) = v11;
-    v12 = (v10 - 8LL);
+    v8 = ((void *)(((int64_t)(v6))));
+    v9 = ((void *)((((int64_t)(v6)) & -16LL)));
+    v10 = ((void *)((((int64_t)(v9)) - 8LL)));
+    *((int64_t *)(((int64_t)(v10)))) = v11;
+    v12 = (((int64_t)(v10)) - 8LL);
     *((int64_t *)(v12)) = v12;
     v14 = (v13 ^ v13);
     v16 = (v15 ^ v15);
     v18 = (v17 + 4128LL);
     v19 = v18;
-    v20 = ((long long (*)(long long, long long, long long, long long, long long, long long))(/* opaque: indirect-call */ 0))(v19, v7, v8, v16, v14, v3);
+    v20 = ((long long (*)(long long, long long, long long, long long, long long, long long))(/* opaque: indirect-call */ 0))(v19, v7, ((int64_t)(v8)), v16, v14, v3);
     (/* opaque: hlt */ 0);
     /* structurally unreachable: block 1 */
     __builtin_unreachable();
@@ -164,7 +164,7 @@ int64_t deregister_tm_clones(void) {
     int64_t v3 = 0LL;
     int64_t v4 = 0LL;
     int8_t v5 = 0LL;
-    int64_t v6 = 0LL;
+    void * v6 = ((void *)(0LL));
     int64_t v7 = 0LL;
     int64_t v8 = 0LL;
     int64_t v9 = 0LL;
@@ -179,8 +179,8 @@ L0:;
         /* phi v11 <- (bb0: v2) (bb1: v8) */
         return v11;
     } else {
-        v6 = (v0 + 16328LL);
-        v7 = (*((int64_t *)(v6)));
+        v6 = ((void *)((v0 + 16328LL)));
+        v7 = (*((int64_t *)(((int64_t)(v6)))));
         v8 = v7;
         v9 = (v8 & v8);
         v10 = (v9 == 0LL);
@@ -219,7 +219,7 @@ int64_t register_tm_clones(void) {
     int64_t v8 = 0LL;
     int64_t v9 = 0LL;
     int64_t v10 = 0LL;
-    int64_t v11 = 0LL;
+    void * v11 = ((void *)(0LL));
     int64_t v12 = 0LL;
     int64_t v13 = 0LL;
     int64_t v14 = 0LL;
@@ -254,17 +254,17 @@ int64_t register_tm_clones(void) {
 /* switch_tables: 0 */
 int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
     int64_t v0 = 0LL;
-    int64_t v1 = 0LL;
+    void * v1 = ((void *)(0LL));
     int8_t v2 = 0LL;
     int8_t v3 = 0LL;
     int64_t v4 = 0LL;
-    int64_t v5 = 0LL;
+    void * v5 = ((void *)(0LL));
     int64_t v6 = 0LL;
-    int64_t v7 = 0LL;
+    void * v7 = ((void *)(0LL));
     int64_t v8 = 0LL;
-    int64_t v9 = 0LL;
+    void * v9 = ((void *)(0LL));
     int64_t v10 = arg0;
-    int64_t v11 = 0LL;
+    void * v11 = ((void *)(0LL));
     int64_t v12 = 0LL;
     int64_t v13 = 0LL;
     int64_t v14 = arg1;
@@ -275,23 +275,23 @@ int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t 
     int64_t v19 = 0LL;
     int64_t v20 = 0LL;
     int64_t v21 = 0LL;
-    int64_t v22 = 0LL;
+    void * v22 = ((void *)(0LL));
     int64_t v23 = 0LL;
-    int64_t v24 = 0LL;
+    void * v24 = ((void *)(0LL));
     int64_t v25 = 0LL;
     int64_t v26 = 0LL;
 
-    v1 = (v0 + 16400LL);
-    v2 = (*((int8_t *)(v1)));
+    v1 = ((void *)((v0 + 16400LL)));
+    v2 = (*((int8_t *)(((int64_t)(v1)))));
     v3 = (v2 != 0LL);
     if (v3) {
         return v26;
     } else {
-        v5 = (v4 - 8LL);
-        *((int64_t *)(v5)) = v6;
-        v7 = (v0 + 16352LL);
-        v8 = (*((int64_t *)(v7)));
-        v9 = v5;
+        v5 = ((void *)((v4 - 8LL)));
+        *((int64_t *)(((int64_t)(v5)))) = v6;
+        v7 = ((void *)((v0 + 16352LL)));
+        v8 = (*((int64_t *)(((int64_t)(v7)))));
+        v9 = ((void *)(((int64_t)(v5))));
         /* structurally unreachable: block 1 */
         __builtin_unreachable();
     }